BalanceSheet,IncomeState For Calculation

+++ Hotel-backend/Common/Dto/IncomeStateDto.cs
@@ -11,45 +11,45 @@ public class IncomeStateDto
     public int MonthID { get; set; }
     public int QuarterNo { get; set; }
     public int GroupID { get; set; }
-    public int Room1 { get; set; }
-    public int FoodB { get; set; }
-    public int FoodB1 { get; set; }
-    public int FoodB2 { get; set; }
-    public int FoodB3 { get; set; }
-    public int FoodB4 { get; set; }
-    public int FoodB5 { get; set; }
-    public int Other { get; set; }
-    public int Other1 { get; set; }
-    public int Other2 { get; set; }
-    public int Other3 { get; set; }
-    public int Other4 { get; set; }
-    public int Other5 { get; set; }
-    public int Other6 { get; set; }
+    public decimal Room1 { get; set; }
+    public decimal FoodB { get; set; }
+    public decimal FoodB1 { get; set; }
+    public decimal FoodB2 { get; set; }
+    public decimal FoodB3 { get; set; }
+    public decimal FoodB4 { get; set; }
+    public decimal FoodB5 { get; set; }
+    public decimal Other { get; set; }
+    public decimal Other1 { get; set; }
+    public decimal Other2 { get; set; }
+    public decimal Other3 { get; set; }
+    public decimal Other4 { get; set; }
+    public decimal Other5 { get; set; }
+    public decimal Other6 { get; set; }
     public int Rent { get; set; }
-    public int TotReven { get; set; }
+    public decimal TotReven { get; set; }
     public int Room { get; set; }
     public int Food2B { get; set; }
     public int Other7 { get; set; }
-    public int TotExpen { get; set; }
-    public int TotDeptIncom { get; set; }
-    public int UndisExpens1 { get; set; }
-    public int UndisExpens2 { get; set; }
-    public int UndisExpens3 { get; set; }
-    public int UndisExpens4 { get; set; }
-    public int UndisExpens5 { get; set; }
-    public int UndisExpens6 { get; set; }
-    public int GrossProfit { get; set; }
+    public decimal TotExpen { get; set; }
+    public decimal TotDeptIncom { get; set; }
+    public decimal UndisExpens1 { get; set; }
+    public decimal UndisExpens2 { get; set; }
+    public decimal UndisExpens3 { get; set; }
+    public decimal UndisExpens4 { get; set; }
+    public decimal UndisExpens5 { get; set; }
+    public decimal UndisExpens6 { get; set; }
+    public decimal GrossProfit { get; set; }
     public int MgtFee { get; set; }
-    public int IncomBfCharg { get; set; }
+    public decimal IncomBfCharg { get; set; }
     public int Property { get; set; }
     public int Insurance { get; set; }
     public int Interest { get; set; }
     public int PropDepreciationerty { get; set; }
     public int TotCharg { get; set; }
-    public int NetIncomBfTAX { get; set; }
-    public int Replace { get; set; }
-    public int AjstNetIncom { get; set; }
-    public int IncomTAX { get; set; }
-    public int NetIncom { get; set; }
+    public decimal NetIncomBfTAX { get; set; }
+    public decimal Replace { get; set; }
+    public decimal AjstNetIncom { get; set; }
+    public decimal IncomTAX { get; set; }
+    public decimal NetIncom { get; set; }
 }
